[wip][add] add inst_16/C_extension
