+++ src/cpu/execution.rs
@@ -3,8 +3,8 @@ mod inst_32;
 
 use super::CPU;
 use super::instruction::Instruction;
-use exe_inst_16::exe_cinst;
-use exe_inst_32::exe_inst;
+use inst_16::exe_cinst;
+use inst_32::exe_inst;
 
 pub trait Execution {
     fn execution(&self, cpu: &mut CPU);
